requirements.txt
Mastodon.py
markdownify
NekoMimi == 1.3.0